+++ tests/run/pep526_variable_annotations.py
@@ -603,34 +603,6 @@ if sys.version_info >= (3, 15) or cython.compiled:
             print(cython.typeof(b + c))
             print(b + c)
 
-
-class A:
-    pass
-
-
-def test_generator_next_node_coercion(N: list[int]):
-    """
-    >>> test_generator_next_node_coercion([A()])  # doctest:+IGNORE_EXCEPTION_DETAIL
-    Traceback (most recent call last):
-    ...
-    TypeError: '<' not supported between instances of 'A' and 'int'
-    """
-    return any(n < 0 for n in N)
-
-
-def test_iterator_next_node_coercion(N: list[int]):
-    """
-    >>> test_iterator_next_node_coercion([A()])  # doctest:+IGNORE_EXCEPTION_DETAIL
-    Traceback (most recent call last):
-    ...
-    TypeError: '<' not supported between instances of 'A' and 'int'
-    """
-    for n in N:
-        if n < 0:
-            return True
-    return False
-
-
 try:
     import numpy.typing as npt
     import numpy as np
