+++ tests/run/type_inference_generics.py
@@ -1,0 +1,33 @@
+# cython: language_level=3
+# mode: run
+# tag: pure3.7, pep526, pep484, warnings
+
+# for the benefit of the pure tests, don't require annotations
+# to be evaluated
+from __future__ import annotations
+
+class A:
+    pass
+
+
+def test_generator_next_node_coercion(N: list[int]):
+    """
+    >>> test_generator_next_node_coercion([A()])  # doctest:+IGNORE_EXCEPTION_DETAIL
+    Traceback (most recent call last):
+    ...
+    TypeError: '<' not supported between instances of 'A' and 'int'
+    """
+    return any(n < 0 for n in N)
+
+
+def test_iterator_next_node_coercion(N: list[int]):
+    """
+    >>> test_iterator_next_node_coercion([A()])  # doctest:+IGNORE_EXCEPTION_DETAIL
+    Traceback (most recent call last):
+    ...
+    TypeError: '<' not supported between instances of 'A' and 'int'
+    """
+    for n in N:
+        if n < 0:
+            return True
+    return False
